create MVVM patern
