+++ app/src/main/java/islom/din/myapplication/model/ImageMessage.kt
@@ -1,0 +1,5 @@
+package islom.din.myapplication.model
+
+data class ImageMessage(
+    val imageRes: Int
+)
