+++ app/src/main/java/islom/din/myapplication/model/Message.kt
@@ -1,0 +1,8 @@
+package islom.din.myapplication.model
+
+data class Message(
+    val id: Int,
+    val simpleMessage: SimpleMessage?,
+    val imageMessage: ImageMessage?
+)
+
