+++ app/src/main/java/islom/din/myapplication/model/SimpleMessage.kt
@@ -1,0 +1,5 @@
+package islom.din.myapplication.model
+
+data class SimpleMessage(
+    val message: String
+)
